+++ customer_portal/src/epilot/models/operations/getallorders.py
@@ -20,8 +20,8 @@ class GetAllOrdersSecurity:
 
 @dataclasses.dataclass
 class GetAllOrdersRequest:
-    from_: Optional[float] = dataclasses.field(default=None, metadata={'query_param': { 'field_name': 'from', 'style': 'form', 'explode': True }})
-    size: Optional[float] = dataclasses.field(default=None, metadata={'query_param': { 'field_name': 'size', 'style': 'form', 'explode': True }})
+    from_: Optional[float] = dataclasses.field(default=0, metadata={'query_param': { 'field_name': 'from', 'style': 'form', 'explode': True }})
+    size: Optional[float] = dataclasses.field(default=100, metadata={'query_param': { 'field_name': 'size', 'style': 'form', 'explode': True }})